pyproject.toml
[tool.poetry]
name = "chatburo"
version = "6.6.6"
description = ""
authors = ["sloev <johanned.valbjorn@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
sanic = "^20.9.0"
trustpilot-json-logging = ">=1.0.12"
PyJWT = "^1.7.1"
aiosqlite = "^0.15.0"
pillow = "^7.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
taskipy = "^1.3.0"

[tool.poetry.scripts]
start = 'chatburo.app:run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

